[networks] Separate HTTP eBPF code from tracer code
