+++ omnibus/config/software/datadog-agent-finalize.rb
@@ -192,6 +192,7 @@ build do
             # Do not strip eBPF programs
             strip_exclude("*tracer*")
             strip_exclude("*offset-guess*")
+            strip_exclude("*http*")
             strip_exclude("*runtime-security*")
         end
