+++ pkg/ebpf/bytecode/runtime/tracer.go
@@ -3,4 +3,4 @@
 
 package runtime
 
-var Tracer = NewRuntimeAsset("tracer.c", "d064b8cc7856937fedda1c65bd36b44e0ba32efdbbb7031fa7f1b4f24d63d61f")
+var Tracer = NewRuntimeAsset("tracer.c", "ed3009e10626138fbbe493c4c8bd7040361a4e46f9ec3e356999610a7bb0d16b")
