+++ pkg/network/ebpf/manager.go
@@ -34,7 +34,7 @@ func NewOffsetManager() *manager.Manager {
 	}
 }
 
-func NewManager(closedHandler, httpHandler *ebpf.PerfHandler, runtimeTracer bool) *manager.Manager {
+func NewManager(closedHandler *ebpf.PerfHandler, runtimeTracer bool) *manager.Manager {
 	mgr := &manager.Manager{
 		Maps: []*manager.Map{
 			{Name: string(probes.ConnMap)},
@@ -45,9 +45,6 @@ func NewManager(closedHandler, httpHandler *ebpf.PerfHandler, runtimeTracer bool
 			{Name: string(probes.UdpPortBindingsMap)},
 			{Name: "pending_bind"},
 			{Name: string(probes.TelemetryMap)},
-			{Name: string(probes.HttpInFlightMap)},
-			{Name: string(probes.HttpBatchesMap)},
-			{Name: string(probes.HttpBatchStateMap)},
 		},
 		PerfMaps: []*manager.PerfMap{
 			{
@@ -59,20 +56,10 @@ func NewManager(closedHandler, httpHandler *ebpf.PerfHandler, runtimeTracer bool
 					LostHandler:        closedHandler.LostHandler,
 				},
 			},
-			{
-				Map: manager.Map{Name: string(probes.HttpNotificationsMap)},
-				PerfMapOptions: manager.PerfMapOptions{
-					PerfRingBufferSize: 8 * os.Getpagesize(),
-					Watermark:          1,
-					DataHandler:        httpHandler.DataHandler,
-					LostHandler:        httpHandler.LostHandler,
-				},
-			},
 		},
 		Probes: []*manager.Probe{
 			{Section: string(probes.TCPSendMsg)},
 			{Section: string(probes.TCPSendMsgPre410), MatchFuncName: "^tcp_sendmsg$"},
-			{Section: string(probes.TCPSendMsgReturn), KProbeMaxActive: maxActive},
 			{Section: string(probes.TCPCleanupRBuf)},
 			{Section: string(probes.TCPClose)},
 			{Section: string(probes.TCPCloseReturn), KProbeMaxActive: maxActive},
@@ -92,7 +79,6 @@ func NewManager(closedHandler, httpHandler *ebpf.PerfHandler, runtimeTracer bool
 			{Section: string(probes.InetBindRet), KProbeMaxActive: maxActive},
 			{Section: string(probes.Inet6BindRet), KProbeMaxActive: maxActive},
 			{Section: string(probes.SocketDnsFilter)},
-			{Section: string(probes.SocketHTTPFilter)},
 			{Section: string(probes.IPRouteOutputFlow)},
 			{Section: string(probes.IPRouteOutputFlowReturn), KProbeMaxActive: maxActive},
 		},
