+++ pkg/network/http/batch_manager.go
@@ -12,7 +12,7 @@ import (
 )
 
 /*
-#include "../ebpf/c/tracer.h"
+#include "../ebpf/c/http-types.h"
 */
 import "C"
