+++ pkg/network/http/monitor.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
+	"github.com/DataDog/datadog-agent/pkg/network/config"
 	"github.com/DataDog/datadog-agent/pkg/network/ebpf/probes"
 	filterpkg "github.com/DataDog/datadog-agent/pkg/network/filter"
 	"github.com/DataDog/ebpf/manager"
@@ -24,6 +25,7 @@ import (
 type Monitor struct {
 	handler func([]httpTX)
 
+	ebpfProgram  *ebpfProgram
 	batchManager *batchManager
 	perfMap      *manager.PerfMap
 	perfHandler  *ddebpf.PerfHandler
@@ -39,13 +41,22 @@ type Monitor struct {
 }
 
 // NewMonitor returns a new Monitor instance
-func NewMonitor(procRoot string, maxEntries int, mgr *manager.Manager, h *ddebpf.PerfHandler) (*Monitor, error) {
+func NewMonitor(c *config.Config) (*Monitor, error) {
+	mgr, err := newEBPFProgram(c)
+	if err != nil {
+		return nil, fmt.Errorf("error setting up http ebpf program: %s", err)
+	}
+
 	filter, _ := mgr.GetProbe(manager.ProbeIdentificationPair{Section: string(probes.SocketHTTPFilter)})
 	if filter == nil {
 		return nil, fmt.Errorf("error retrieving socket filter")
 	}
 
-	closeFilterFn, err := filterpkg.HeadlessSocketFilter(procRoot, filter)
+	if err := mgr.Init(); err != nil {
+		return nil, fmt.Errorf("error initializing http ebpf program: %s", err)
+	}
+
+	closeFilterFn, err := filterpkg.HeadlessSocketFilter(c.ProcRoot, filter)
 	if err != nil {
 		return nil, fmt.Errorf("error enabling HTTP traffic inspection: %s", err)
 	}
@@ -69,7 +80,7 @@ func NewMonitor(procRoot string, maxEntries int, mgr *manager.Manager, h *ddebpf
 	}
 
 	telemetry := newTelemetry()
-	statkeeper := newHTTPStatkeeper(maxEntries, telemetry)
+	statkeeper := newHTTPStatkeeper(c.MaxHTTPStatsBuffered, telemetry)
 
 	handler := func(transactions []httpTX) {
 		if statkeeper != nil {
@@ -79,9 +90,10 @@ func NewMonitor(procRoot string, maxEntries int, mgr *manager.Manager, h *ddebpf
 
 	return &Monitor{
 		handler:       handler,
+		ebpfProgram:   mgr,
 		batchManager:  newBatchManager(batchMap, batchStateMap, numCPUs),
 		perfMap:       pm,
-		perfHandler:   h,
+		perfHandler:   mgr.perfHandler,
 		telemetry:     telemetry,
 		pollRequests:  make(chan chan map[Key]RequestStats),
 		closeFilterFn: closeFilterFn,
@@ -174,6 +186,7 @@ func (m *Monitor) Stop() {
 		return
 	}
 
+	m.ebpfProgram.Stop(manager.CleanAll)
 	m.closeFilterFn()
 	_ = m.perfMap.Stop(manager.CleanAll)
 	m.perfHandler.Stop()
